pimcore 3.0.6

+++ lib/Pimcore/Update.php
@@ -193,6 +193,7 @@ class Update {
         $db = Resource::get();
         
         $db->query("CREATE TABLE IF NOT EXISTS `" . self::$tmpTable . "` (
+          `id` int(11) NULL DEFAULT NULL,
           `revision` int(11) NULL DEFAULT NULL,
           `path` varchar(255) NULL DEFAULT NULL,
           `action` varchar(50) NULL DEFAULT NULL
@@ -213,22 +214,25 @@ class Update {
             File::mkdir($scriptsDir);
         }
         
-        
         $xml = Tool::getHttpData($url);
         if($xml) {
-            $parserOptions = LIBXML_VERSION >= 20900 ? LIBXML_NOCDATA | LIBXML_PARSEHUGE : LIBXML_NOCDATA;
+            $parserOptions = LIBXML_NOCDATA;
+            if(defined("LIBXML_PARSEHUGE")) {
+                $parserOptions = LIBXML_NOCDATA | LIBXML_PARSEHUGE;
+            }
+
             $updateFiles = simplexml_load_string($xml, null, $parserOptions);
             
             foreach ($updateFiles->file as $file) {
                 
                 if($file->type == "file") {
                     if ($file->action == "update" || $file->action == "add") {
-                        $newPath = str_replace("/","~~~",$file->path);
-                        $newFile = $filesDir."/".$newPath;
+                        $newFile = $filesDir . "/" . $file->id . "-" . $file->revision;
                         File::put($newFile, base64_decode((string) $file->content));
                     }
                     
                     $db->insert(self::$tmpTable, array(
+                        "id" => $file->id,
                         "revision" => $revision,
                         "path" => (string) $file->path,
                         "action" => (string)$file->action
@@ -256,7 +260,16 @@ class Update {
                         File::mkdir(dirname(PIMCORE_DOCUMENT_ROOT . $file["path"]));
                     }
                 }
-                $srcFile = PIMCORE_SYSTEM_TEMP_DIRECTORY . "/update/".$revision."/files/" . str_replace("/","~~~",$file["path"]);
+
+                if(array_key_exists("id", $file) && $file["id"]) {
+                    // this is the new style, see https://www.pimcore.org/issues/browse/PIMCORE-2722
+                    $srcFile = PIMCORE_SYSTEM_TEMP_DIRECTORY . "/update/".$revision."/files/" . $file["id"] . "-" . $file["revision"];
+                } else {
+                    // this is the old style, which we still have to support here, otherwise there's the risk that the
+                    // running update cannot be finished
+                    $srcFile = PIMCORE_SYSTEM_TEMP_DIRECTORY . "/update/".$revision."/files/" . str_replace("/","~~~",$file["path"]);
+                }
+
                 $destFile = PIMCORE_DOCUMENT_ROOT . $file["path"];
                 
                 if(!self::$dryRun) {
@@ -265,7 +278,9 @@ class Update {
             }
             else if ($file["action"] == "delete") {
                 if(!self::$dryRun) {
-                    unlink(PIMCORE_DOCUMENT_ROOT . $file["path"]);
+                    if (file_exists(PIMCORE_DOCUMENT_ROOT . $file["path"])) {
+                        unlink(PIMCORE_DOCUMENT_ROOT . $file["path"]);
+                    }
         
                     // remove also directory if its empty
                     if (count(glob(dirname(PIMCORE_DOCUMENT_ROOT . $file["path"]) . "/*")) === 0) {
